requirement.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anthropic==0.36.2
antlr4-python3-runtime==4.11.0
anyio==4.7.0
astor==0.8.1
attrs==24.3.0
av==14.0.1
beautifulsoup4==4.12.3
blake3==1.0.1
bleach==6.2.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cmudict==1.0.32
colorama==0.4.6
colorful==0.5.6
compressed-tensors==0.8.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dataframe-image==0.2.6
datasets==3.2.0
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
einops==0.8.0
evaluate==0.4.3
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flash-attn==2.7.2.post1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
geographiclib==2.0
geopy==2.4.1
gguf==0.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html2image==2.0.5
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.5.2
iniconfig==2.0.0
interegular==0.3.3
jieba==0.42.1
jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.8
lark==1.2.2
linkify-it-py==2.0.3
lm-format-enforcer==0.10.9
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral-common==1.5.1
mistune==3.1.0
more-itertools==10.5.0
mosestokenizer==1.0.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.59.3
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openfile==0.0.7
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
partial-json-parser==0.2.1.1.post4
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
pronouncing==0.2.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.4
pydantic-core==2.27.2
pyext==0.5
pygments==2.18.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
rapidfuzz==3.9.5
ray==2.40.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-cache==1.2.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.4.3
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.21.0
toolwrapper==2.1.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
typing-extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unidecode==1.3.8
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
vllm==0.6.6.post1
watchfiles==1.0.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
xformers==0.0.28.post3
xgrammar==0.1.8
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
